requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
pillow>=9.0.0
piexif>=1.1.3
faker>=18.0.0
pyotp>=2.8.0
yt-dlp>=2023.1.1
requests>=2.28.0
